requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
confusable-homoglyphs==3.2.0
Django==4.0.4
django-registration==3.2
djangorestframework==3.13.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
